Valoren mother-fuckers, me explota la cabeza

+++ src/controllers/admin/removeProduct.controller.js
@@ -20,6 +20,11 @@
 
 const db = require("../../database/models");
 module.exports = (req, res) => {
+  const userLogin = res.locals.userLogin;
+
+  if (!userLogin) {
+      return res.redirect('aut/login'); 
+  }
   const { id } = req.params;
   db.Product.destroy({
     where: {
